requirements.txt
aiohttp==3.9.1
aiortc==1.6.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
python-dotenv==1.0.0
gtts==2.3.2
speechrecognition==3.10.0
pydub==0.25.1
moviepy==1.0.3
websockets==12.0
python-multipart==0.0.6